[AIRFLOW-1314] Add support for volume mounts & Secrets
