+++ airflow/contrib/operators/k8s_pod_operator/k8s_pod_operator.py
@@ -18,74 +18,61 @@ from airflow.exceptions import AirflowException
 from airflow.operators.python_operator import PythonOperator
 from airflow.utils.decorators import apply_defaults
 from airflow.contrib.kubernetes.pod_launcher import KubernetesLauncher, \
-    KubernetesCommunicationService
+    KubernetesCommunicationService, incluster_namespace
 from airflow.contrib.kubernetes.kubernetes_request_factory import \
     SimplePodRequestFactory, \
     ReturnValuePodRequestFactory
-from .op_context import OpContext
 
 
 class PodOperator(PythonOperator):
     """
         Executes a pod and waits for the job to finish.
+
         :param dag_run_id: The unique run ID that would be attached to the pod as a label
         :type dag_run_id: str
         :param pod_factory: Reference to the function that creates the pod with format:
                             function (OpContext) => Pod
         :type pod_factory: callable
-        :param cache_output: If set to true, the output of the pod would be saved in a
-                            cache object using md5 hash of all the pod parameters
-                            and in case of success, the cached results will be returned
-                            on consecutive calls. Only use this
     """
     # template_fields = tuple('dag_run_id')
     ui_color = '#8da7be'
 
     @apply_defaults
     def __init__(
-        self,
-        dag_run_id,
-        pod_factory,
-        cache_output,
-        kube_request_factory=None,
-        *args,
-        **kwargs
-    ):
-        super(PodOperator, self).__init__(
-            python_callable=lambda _: 1,
-            provide_context=True,
-            *args,
-            **kwargs)
+            self,
+            dag_run_id,
+            pod_factory,
+            kube_request_factory=None,
+            *args, **kwargs):
+        super(PodOperator, self).__init__(python_callable=lambda _: 1, provide_context=True, *args, **kwargs)
         self.logger = logging.getLogger(self.__class__.__name__)
         if not callable(pod_factory):
             raise AirflowException('`pod_factory` param must be callable')
         self.dag_run_id = dag_run_id
         self.pod_factory = pod_factory
-        self._cache_output = cache_output
-        self.op_context = OpContext(self.task_id)
         self.kwargs = kwargs
         self._kube_request_factory = kube_request_factory or SimplePodRequestFactory
 
     def execute(self, context):
-        task_instance = context.get('task_instance')
-        if task_instance is None:
-            raise AirflowException('`task_instance` is empty! This should not happen')
-        self.op_context.set_xcom_instance(task_instance)
-        pod = self.pod_factory(self.op_context, context)
+        pod = self.get_pod_object(context)
+
         # Customize the pod
         pod.name = self.task_id
         pod.labels['run_id'] = self.dag_run_id
-        pod.namespace = self.dag.default_args.get('namespace', pod.namespace)
+        try:
+            pod.namespace = self.dag.default_args.get('namespace', pod.namespace) or incluster_namespace()
+        except:
+            # Used default namespace
+            pass
 
         # Launch the pod and wait for it to finish
         KubernetesLauncher(pod, self._kube_request_factory).launch()
-        self.op_context.result = pod.result
+        result = pod.result
+        context['ti'].xcom_push(key='result', value=result)
 
-        # Cache the output
         custom_return_value = self.on_pod_success(context)
-        if custom_return_value:
-            self.op_context.custom_return_value = custom_return_value
-        return self.op_context.result
+        self.set_custom_return_value(context, custom_return_value)
+        return result
 
     def on_pod_success(self, context):
         """
@@ -95,32 +82,39 @@ class PodOperator(PythonOperator):
         """
         pass
 
+    def get_pod_object(self, context):
+        """
+            Returns a pod object. Overwrite this method to define custom objects
+        :param context: The task context
+        :return: The pod object
+        """
+        return self.pod_factory(context)
+
+    def set_custom_return_value(self, context, custom_return_value):
+        if custom_return_value:
+            context['ti'].xcom_push(key='custom_result', value=custom_return_value)
+
 
 class ReturnValuePodOperator(PodOperator):
     """
      This pod operators is a normal pod operator with the addition of
      reading custom return value back from kubernetes.
     """
-
     def __init__(self,
-                 kube_com_service_factory,
                  result_data_file,
+                 kube_com_service_factory=None,
                  *args, **kwargs):
         super(ReturnValuePodOperator, self).__init__(*args, **kwargs)
+        kube_com_service_factory = kube_com_service_factory or (
+                 lambda: KubernetesCommunicationService.from_dag_default_args(self.dag))
         if not isinstance(kube_com_service_factory(), KubernetesCommunicationService):
-            raise AirflowException(
-                '`kube_com_service_factory` must be of type '
-                'KubernetesCommunicationService')
+            raise AirflowException('`kube_com_service_factory` must be of type KubernetesCommunicationService')
         self._kube_com_service_factory = kube_com_service_factory
         self._result_data_file = result_data_file
-        self._kube_request_factory = self._return_value_kube_request  # Overwrite the
-        # default request factory
+        self._kube_request_factory = self._return_value_kube_request  # Overwrite the default request factory
 
     def on_pod_success(self, context):
-        return_val = self._kube_com_service_factory().pod_return_data(self.task_id)
-        self.op_context.result = return_val  # We also overwrite the results
-        return return_val
+        return self._kube_com_service_factory().pod_return_data(self.task_id)
 
     def _return_value_kube_request(self):
-        return ReturnValuePodRequestFactory(self._kube_com_service_factory,
-                                            self._result_data_file)
+        return ReturnValuePodRequestFactory(self._kube_com_service_factory, self._result_data_file)
